extracts test helper
